+++ src/main/java/javax/util/streamex/LongStreamEx.java
@@ -23,6 +23,7 @@ import java.util.LongSummaryStatistics;
 import java.util.OptionalDouble;
 import java.util.OptionalLong;
 import java.util.Random;
+import java.util.Map.Entry;
 import java.util.PrimitiveIterator.OfLong;
 import java.util.concurrent.ForkJoinPool;
 import java.util.function.BiConsumer;
@@ -176,7 +177,27 @@ public class LongStreamEx implements LongStream {
         return strategy().newDoubleStreamEx(stream.mapToDouble(mapper));
     }
 
-    public <K, V> EntryStream<K, V> mapToEntry(LongFunction<? extends K> keyMapper, LongFunction<? extends V> valueMapper) {
+    /**
+     * Returns an {@link EntryStream} consisting of the {@link Entry} objects
+     * which keys and values are results of applying the given functions to the
+     * elements of this stream.
+     *
+     * <p>
+     * This is an intermediate operation.
+     *
+     * @param <K>
+     *            The {@code Entry} key type
+     * @param <V>
+     *            The {@code Entry} value type
+     * @param keyMapper
+     *            a non-interfering, stateless function to apply to each element
+     * @param valueMapper
+     *            a non-interfering, stateless function to apply to each element
+     * @return the new stream
+     * @since 0.3.1
+     */
+    public <K, V> EntryStream<K, V> mapToEntry(LongFunction<? extends K> keyMapper,
+            LongFunction<? extends V> valueMapper) {
         return strategy().newEntryStream(
                 stream.mapToObj(t -> new AbstractMap.SimpleImmutableEntry<>(keyMapper.apply(t), valueMapper.apply(t))));
     }
@@ -337,8 +358,9 @@ public class LongStreamEx implements LongStream {
      * updating the state of the result rather than by replacing the result.
      *
      * <p>
-     * Like {@link #reduce(long, LongBinaryOperator)}, {@code collect} operations
-     * can be parallelized without requiring additional synchronization.
+     * Like {@link #reduce(long, LongBinaryOperator)}, {@code collect}
+     * operations can be parallelized without requiring additional
+     * synchronization.
      *
      * <p>
      * This is a terminal operation.
@@ -907,7 +929,7 @@ public class LongStreamEx implements LongStream {
      * to the every adjacent pair of elements of this stream.
      *
      * <p>
-     * This is an intermediate operation.
+     * This is a quasi-intermediate operation.
      * 
      * <p>
      * The output stream will contain one element less than this stream. If this
@@ -926,14 +948,54 @@ public class LongStreamEx implements LongStream {
                         stream.isParallel()).onClose(stream::close));
     }
 
+    /**
+     * Returns a {@link String} which contains the results of calling
+     * {@link String#valueOf(long)} on each element of this stream, separated by
+     * the specified delimiter, in encounter order.
+     *
+     * <p>
+     * This is a terminal operation.
+     * 
+     * @param delimiter
+     *            the delimiter to be used between each element
+     * @return a {@code String}. For empty input stream empty String is
+     *         returned.
+     * @since 0.3.1
+     */
     public String joining(CharSequence delimiter) {
         return collect(LongCollector.joining(delimiter));
     }
 
+    /**
+     * Returns a {@link String} which contains the results of calling
+     * {@link String#valueOf(long)} on each element of this stream, separated
+     * by the specified delimiter, with the specified prefix and suffix in
+     * encounter order.
+     *
+     * <p>
+     * This is a terminal operation.
+     * 
+     * @param delimiter
+     *            the delimiter to be used between each element
+     * @param prefix
+     *            the sequence of characters to be used at the beginning of the
+     *            joined result
+     * @param suffix
+     *            the sequence of characters to be used at the end of the joined
+     *            result
+     * @return a {@code String}. For empty input stream empty String is
+     *         returned.
+     * @since 0.3.1
+     */
     public String joining(CharSequence delimiter, CharSequence prefix, CharSequence suffix) {
         return collect(LongCollector.joining(delimiter, prefix, suffix));
     }
-    
+
+    /**
+     * Returns an empty sequential {@code LongStreamEx}.
+     *
+     * @return an empty sequential stream
+     */
     public static LongStreamEx empty() {
         return new LongStreamEx(LongStream.empty());
     }