+++ src/main/java/javax/util/streamex/package-info.java
@@ -31,6 +31,15 @@
  * <p>
  * {@link javax.util.streamex.IntCollector}, {@link javax.util.streamex.LongCollector}, {@link javax.util.streamex.DoubleCollector}: specialized collectors to work efficiently with primitive streams.
  * 
+ * <p>
+ * Most of new stream operations are either intermediate or terminal like it's defined in Java 8 Stream API. However due to the API limitations
+ * a few operations are called "quasi-intermediate". In the most of the cases they behave as intermediate operations: for sequential stream there 
+ * should be no visible difference between intermediate and quasi-intermediate operation. The only known difference
+ * is handling a parallel stream status. For intermediate operation there's no difference on calling {@code parallel()} 
+ * before or after any intermediate operation. For quasi-intermediate operations if you call {@code parallel()} after the operation, then previous
+ * steps will remain sequential. Similarly if you create a parallel stream, perform some intermediate operations, use quasi-intermediate operation,
+ * then call {@code sequential()}, the steps before quasi-intermediate operation may still be executed in parallel. 
+ * 
  * @author Tagir Valeev
  */
 package javax.util.streamex;