+++ src/test/java/javax/util/streamex/LongStreamExTest.java
@@ -16,6 +16,8 @@
 package javax.util.streamex;
 
 import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
 import java.util.OptionalLong;
 import java.util.Random;
 import java.util.Spliterator;
@@ -177,5 +179,20 @@ public class LongStreamExTest {
                 .append(LongStream.empty()).parallel().pairMap((a, b) -> b - a).toArray());
         assertArrayEquals(LongStreamEx.range(1, 100).toArray(), LongStreamEx.range(100).map(i -> i * (i + 1) / 2)
                 .prepend(LongStream.empty()).parallel().pairMap((a, b) -> b - a).toArray());
+    }
+    
+    @Test
+    public void testJoining() {
+        assertEquals("0,1,2,3,4,5,6,7,8,9", LongStreamEx.range(10).joining(","));
+        assertEquals("0,1,2,3,4,5,6,7,8,9", LongStreamEx.range(10).parallel().joining(","));
+        assertEquals("[0,1,2,3,4,5,6,7,8,9]", LongStreamEx.range(10).joining(",", "[", "]"));
+        assertEquals("[0,1,2,3,4,5,6,7,8,9]", LongStreamEx.range(10).parallel().joining(",", "[", "]"));
+    }
+    
+    @Test
+    public void testMapToEntry() {
+        Map<Long, List<Long>> result = LongStreamEx.range(10).mapToEntry(x -> x % 2, x -> x).grouping();
+        assertEquals(Arrays.asList(0l, 2l, 4l, 6l, 8l), result.get(0l));
+        assertEquals(Arrays.asList(1l, 3l, 5l, 7l, 9l), result.get(1l));
     }
 }