+++ CHANGES.md
@@ -1,5 +1,14 @@
 # StreamEx changes
 
+### 0.3.1
+
+* Added mapToEntry method for primitive streams
+* Added joining methods family for primitive streams
+* Added StreamEx.collapse/groupRuns methods
+* Added StreamEx.distinct(atLeast) method
+* Released jar works now as an OSGi bundle
+* Updated documentation
+
 ### 0.3.0
 
 * Added primitive collectors: IntCollector, LongCollector, DoubleCollector